// mcp-servers/gatling/src/apiClientGenerated/gatlingEnterpriseSchemas.ts
export type ManagedLocationType = "managed_location";

export type PackageType = "package";

export type PrivateLocationType = "private_location";
// ...
export type ArtifactFormatEnum = "jvm" | "js";

export type LocationCapabilitiesDto = {
  artifactFormats: ArtifactFormatEnum[];
};

export type LocationItemResponse = PrivateLocationItemResponse | ManagedLocationItemResponse;

export type LocationReadAllResponse = {
  data: LocationItemResponse[];
};

export type LocationTypeEnum = "private_location" | "managed_location";

export type ManagedLocationItemResponse = {
  /**
   * @default managed_location
   */
  _type: ManagedLocationType & void;
  /**
   * Location ID
   */
  _id: string;
  /**
   * Location name
   */
  _name: string;
  _capabilities: LocationCapabilitiesDto;
};

export type PrivateLocationItemResponse = {
  /**
   * @default private_location
   */
  _type: PrivateLocationType & void;
  /**
   * Location ID
   */
  _id: string;
  /**
   * Non-empty trimmed string
   */
  _description?: string;
  _capabilities: LocationCapabilitiesDto;
};